+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/StandaloneScenarioExecutor.java
@@ -13,8 +13,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import net.sf.cglib.proxy.Enhancer;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,29 +20,20 @@ import com.google.common.base.Optional;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.tngtech.jgiven.CurrentStep;
-import com.tngtech.jgiven.annotation.AfterScenario;
-import com.tngtech.jgiven.annotation.AfterStage;
-import com.tngtech.jgiven.annotation.BeforeScenario;
-import com.tngtech.jgiven.annotation.BeforeStage;
-import com.tngtech.jgiven.annotation.Hidden;
-import com.tngtech.jgiven.annotation.NotImplementedYet;
-import com.tngtech.jgiven.annotation.ScenarioRule;
-import com.tngtech.jgiven.annotation.ScenarioStage;
+import com.tngtech.jgiven.annotation.*;
 import com.tngtech.jgiven.attachment.Attachment;
 import com.tngtech.jgiven.exception.FailIfPassedException;
 import com.tngtech.jgiven.exception.JGivenUserException;
 import com.tngtech.jgiven.impl.inject.ValueInjector;
-import com.tngtech.jgiven.impl.intercept.InvocationMode;
-import com.tngtech.jgiven.impl.intercept.NoOpScenarioListener;
-import com.tngtech.jgiven.impl.intercept.ScenarioListener;
-import com.tngtech.jgiven.impl.intercept.StandaloneStepMethodInterceptor;
-import com.tngtech.jgiven.impl.intercept.StepMethodHandler;
+import com.tngtech.jgiven.impl.intercept.*;
 import com.tngtech.jgiven.impl.util.FieldCache;
 import com.tngtech.jgiven.impl.util.ParameterNameUtil;
 import com.tngtech.jgiven.impl.util.ReflectionUtil;
 import com.tngtech.jgiven.impl.util.ReflectionUtil.MethodAction;
 import com.tngtech.jgiven.integration.CanWire;
 import com.tngtech.jgiven.report.model.NamedArgument;
+
+import net.sf.cglib.proxy.Enhancer;
 
 /**
  * Main class of JGiven for executing scenarios.
@@ -447,7 +436,17 @@ public class StandaloneScenarioExecutor implements ScenarioExecutor {
     public void startScenario( Method method, List<NamedArgument> arguments ) {
         listener.scenarioStarted( method, arguments );
 
-        if( method.isAnnotationPresent( NotImplementedYet.class ) ) {
+        if( method.isAnnotationPresent( Pending.class ) ) {
+            Pending annotation = method.getAnnotation( Pending.class );
+
+            if( annotation.failIfPass() ) {
+                failIfPass();
+            } else if( !annotation.executeSteps() ) {
+                methodInterceptor.disableMethodExecution();
+                executeLifeCycleMethods = false;
+            }
+            suppressExceptions = true;
+        } else if( method.isAnnotationPresent( NotImplementedYet.class ) ) {
             NotImplementedYet annotation = method.getAnnotation( NotImplementedYet.class );
 
             if( annotation.failIfPass() ) {
@@ -458,6 +457,7 @@ public class StandaloneScenarioExecutor implements ScenarioExecutor {
             }
             suppressExceptions = true;
         }
+
     }
 
     @Override