+++ jgiven-examples/src/test/java/com/tngtech/jgiven/examples/notimplementedyet/PendingExampleTest.java
@@ -3,29 +3,29 @@ package com.tngtech.jgiven.examples.notimplementedyet;
 import org.junit.Test;
 
 import com.tngtech.jgiven.annotation.Description;
-import com.tngtech.jgiven.annotation.NotImplementedYet;
+import com.tngtech.jgiven.annotation.Pending;
 import com.tngtech.jgiven.junit.SimpleScenarioTest;
-import com.tngtech.jgiven.tags.FeatureNotImplementedYet;
+import com.tngtech.jgiven.tags.FeaturePending;
 
 @Description( "As a good BDD practitioner,<br>"
         + "I want to write my scenarios before I start coding<br>"
         + "In order to discuss them with business stakeholders" )
-public class NotImplementYetExampleTest extends SimpleScenarioTest<NotImplementYetExampleTest.TestSteps> {
+public class PendingExampleTest extends SimpleScenarioTest<PendingExampleTest.TestSteps> {
 
     @Test
-    @FeatureNotImplementedYet
-    @NotImplementedYet
-    public void scenarios_that_are_not_implemented_yet_can_be_annotated_with_the_NotImplementedYet_annotation() {
+    @FeaturePending
+    @Pending
+    public void scenarios_that_are_pending_can_be_annotated_with_the_Pending_annotation() {
         given().some_state();
         when().some_action();
         then().some_result();
     }
 
     @Test
-    @FeatureNotImplementedYet
-    public void single_steps_can_be_annotated_with_NotImplementedYet() {
+    @FeaturePending
+    public void single_steps_can_be_annotated_with_Pending() {
         given().some_state();
-        when().some_not_implemented_yet_action();
+        when().some_pending_action();
         then().some_result();
     }
 
@@ -43,8 +43,8 @@ public class NotImplementYetExampleTest extends SimpleScenarioTest<NotImplementY
             return this;
         }
 
-        @NotImplementedYet
-        public TestSteps some_not_implemented_yet_action() {
+        @Pending
+        public TestSteps some_pending_action() {
             return this;
         }
     }