+++ jgiven-tests/src/main/java/com/tngtech/jgiven/tags/FeaturePending.java
@@ -6,11 +6,11 @@ import java.lang.annotation.RetentionPolicy;
 import com.tngtech.jgiven.annotation.IsTag;
 
 @FeatureCore
-@IsTag( name = "NotImplementedYet Annotation",
+@IsTag( name = "Pending Annotation",
     description = "As a good BDD practitioner,<br>"
             + "I want to write my scenarios before I start coding<br>"
             + "In order to discuss them with business stakeholders" )
 @Retention( RetentionPolicy.RUNTIME )
-public @interface FeatureNotImplementedYet {
+public @interface FeaturePending {
 
 }